backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
future==0.18.2
identify==2.2.10
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
nodeenv==1.6.0
openapi-codec==1.3.2
packaging==21.0
pre-commit==2.13.0
pyparsing==2.4.7
python-whois==0.7.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.7
